wip: complete but issues with four key bindings
<, >, ? and : are not appearing, likely due to shifted() and Shift
modifier use when looking up key.
