+++ src/components/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod prompt;
 pub(crate) mod editor;
 pub(crate) mod editor_keymap;
 mod editor_keymap_legend;
+pub(crate) mod editor_keymap_printer;
 pub(crate) mod file_explorer;
 pub(crate) mod keymap_legend;
 pub(crate) mod render_editor;
